+++ tests/Executor/NonNullTest.php
@@ -550,7 +550,7 @@ class NonNullTest extends TestCase
             'data'   => ['syncNest' => null],
             'errors' => [
                 [
-                    'debugMessage' => 'Cannot return null for non-nullable field DataType.syncNonNull.',
+                    'debugMessage' => 'Cannot return null for non-nullable field "DataType.syncNonNull".',
                     'locations'    => [['line' => 4, 'column' => 11]],
                 ],
             ],
@@ -577,7 +577,7 @@ class NonNullTest extends TestCase
             'data'   => ['syncNest' => null],
             'errors' => [
                 [
-                    'debugMessage' => 'Cannot return null for non-nullable field DataType.promiseNonNull.',
+                    'debugMessage' => 'Cannot return null for non-nullable field "DataType.promiseNonNull".',
                     'locations'    => [['line' => 4, 'column' => 11]],
                 ],
             ],
@@ -605,7 +605,7 @@ class NonNullTest extends TestCase
             'data'   => ['promiseNest' => null],
             'errors' => [
                 [
-                    'debugMessage' => 'Cannot return null for non-nullable field DataType.syncNonNull.',
+                    'debugMessage' => 'Cannot return null for non-nullable field "DataType.syncNonNull".',
                     'locations'    => [['line' => 4, 'column' => 11]],
                 ],
             ],
@@ -633,7 +633,7 @@ class NonNullTest extends TestCase
             'data'   => ['promiseNest' => null],
             'errors' => [
                 [
-                    'debugMessage' => 'Cannot return null for non-nullable field DataType.promiseNonNull.',
+                    'debugMessage' => 'Cannot return null for non-nullable field "DataType.promiseNonNull".',
                     'locations'    => [['line' => 4, 'column' => 11]],
                 ],
             ],
@@ -775,10 +775,10 @@ class NonNullTest extends TestCase
                 'anotherPromiseNest' => null,
             ],
             'errors' => [
-                ['debugMessage' => 'Cannot return null for non-nullable field DataType.syncNonNull.', 'locations' => [['line' => 8, 'column' => 19]]],
-                ['debugMessage' => 'Cannot return null for non-nullable field DataType.syncNonNull.', 'locations' => [['line' => 19, 'column' => 19]]],
-                ['debugMessage' => 'Cannot return null for non-nullable field DataType.promiseNonNull.', 'locations' => [['line' => 30, 'column' => 19]]],
-                ['debugMessage' => 'Cannot return null for non-nullable field DataType.promiseNonNull.', 'locations' => [['line' => 41, 'column' => 19]]],
+                ['debugMessage' => 'Cannot return null for non-nullable field "DataType.syncNonNull".', 'locations' => [['line' => 8, 'column' => 19]]],
+                ['debugMessage' => 'Cannot return null for non-nullable field "DataType.syncNonNull".', 'locations' => [['line' => 19, 'column' => 19]]],
+                ['debugMessage' => 'Cannot return null for non-nullable field "DataType.promiseNonNull".', 'locations' => [['line' => 30, 'column' => 19]]],
+                ['debugMessage' => 'Cannot return null for non-nullable field "DataType.promiseNonNull".', 'locations' => [['line' => 41, 'column' => 19]]],
             ],
         ];
 
@@ -1024,7 +1024,7 @@ class NonNullTest extends TestCase
         $expected = [
             'errors' => [
                 [
-                    'debugMessage' => 'Cannot return null for non-nullable field DataType.syncNonNull.',
+                    'debugMessage' => 'Cannot return null for non-nullable field "DataType.syncNonNull".',
                     'locations'    => [['line' => 2, 'column' => 17]],
                 ],
             ],
@@ -1046,7 +1046,7 @@ class NonNullTest extends TestCase
         $expected = [
             'errors' => [
                 [
-                    'debugMessage' => 'Cannot return null for non-nullable field DataType.promiseNonNull.',
+                    'debugMessage' => 'Cannot return null for non-nullable field "DataType.promiseNonNull".',
                     'locations'    => [['line' => 2, 'column' => 17]],
                 ],
             ],